% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{lab_data}
\alias{lab_data}
\title{Get units of a measurement in a nice format for a given variable.}
\usage{
lab_data(feature, units)
}
\arguments{
\item{feature}{A character. The variable for which to generate units.}

\item{units}{A vector of units generated from \code{trackeR} package \code{getUnits()}.}
}
\description{
Get units of a measurement in a nice format for a given variable.
}
