% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plot all selected workouts on an interactive leaflet map.}
\usage{
plot_map(sumX, df = NULL, threshold = TRUE, all_sessions, colour_sessions)
}
\arguments{
\item{sumX}{An object of class \code{trackeRdataSummary}.}

\item{df}{A dataframe (optional). A user can provide a dataframe generated by \code{prepRoute}.}

\item{threshold}{Logical. Should thresholds be applied?}

\item{all_sessions}{A vector of sessions being plotted.}

\item{colour_sessions}{A vector of sessions to be colour to orange.}
}
\description{
Plot all selected workouts on an interactive leaflet map.
}
