% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackeRapp.R
\name{trackeR_app}
\alias{trackeR_app}
\title{Launch the trackeR dashboard}
\usage{
trackeR_app(quiet = FALSE)
}
\arguments{
\item{quiet}{Should Shiny status messages be shown? Defaults to FALSE.}
}
\description{
Launch the trackeR dashboard
}
